requirements.txt
Django
Pillow
pip
asgiref
sqlparse
PyMySQL
celery
eventlet
paddlepaddle
paddlehub